+++ lib/core/labels/index.ts
@@ -10,3 +10,4 @@
 export * from './types'
 export * from './resolve'
 export * from './compliance'
+export * from './approval'